refactorización de sortData, avance con el parametro data
