+++ src/data.js
@@ -3,8 +3,13 @@ export const filterData = (data,condition) => {
     return data[condition]
   // return '';  data= datos sin orden  sortBy= con respecto a cual de los datos se va a ordenar  sortorder= forma de orden(ascendete, descendente, alfabeticamente)
 };
-export const sortData = (/*data,sortBy, sortOrder*/) => {
+export const sortData = (data/*,sortBy, sortOrder*/) => {
   // return '';  data= datos sin orden  sortBy= con respecto a cual de los datos se va a ordenar  sortorder= forma de orden(ascendete, descendente, alfabeticamente)
+  let stringdata = JSON.stringify(data);
+  let stringdataParse = JSON.parse(stringdata);
+  //ordenando objetos por title
+  return stringdataParse.sort(GetSortOrder("title"));
+
 };
 
 export const computeStats = (/*data*/) => {
@@ -20,6 +25,16 @@ FUNCIONES RECOMENDADAS
 -computeStats(data)---cálculos estadísticos básicos para ser mostrados de acuerdo a la data proporcionada.()
 */
 
+function GetSortOrder(prop) {    
+  return function(a, b) {    
+      if (a[prop] > b[prop]) {    
+          return 1;    
+      } else if (a[prop] < b[prop]) {    
+          return -1;    
+      }    
+      return 0;    
+  }    
+}
